refactor(people): one-concept files hold the manual-dedupe hook and the list reads — person.go/organization.go back under the 500-line cap
No behavior change. The manual-policy chokepoint wiring moves to
creatededupe.go (manualDedupePerson/Organization + recordIfReview +
recordNearMatch); the list reads move to person_list.go /
organization_list.go on the relationship_list.go precedent, with the
filter chain extracted per file and the name/where literals named. The
deliberately parallel person/org page reads carry a reasoned dupl
waiver: a generic page-runner would abstract over the two record types
for symmetry alone.
