+++ backend/internal/modules/people/person.go
@@ -7,7 +7,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"strings"
 	"time"
 
 	"github.com/gradionhq/margince/backend/internal/platform/auth"
@@ -19,7 +18,6 @@ import (
 	"github.com/gradionhq/margince/backend/internal/shared/apperrors"
 	"github.com/gradionhq/margince/backend/internal/shared/kernel/ids"
 	"github.com/gradionhq/margince/backend/internal/shared/kernel/principal"
-	"github.com/gradionhq/margince/backend/internal/shared/ports/fieldcatalog"
 )
 
 // DuplicateEmailError carries the existing person for the 409 dedupe
@@ -91,15 +89,7 @@ func (s *Store) CreatePerson(ctx context.Context, in CreatePersonInput) (crmcont
 			return err
 		}
 
-		// PO-F-1 chokepoint, manual policy: the pre-check above already
-		// refused every exact hit in this same transaction (and the
-		// person_email unique index backs it under races), so the
-		// chokepoint's remaining signal is the fuzzy tier. It must run
-		// BEFORE the insert — afterwards the new row would match itself.
-		match, err := DedupePerson(ctx, tx, PersonCandidate{
-			FullName: in.FullName,
-			Emails:   dedupeCandidateEmails(in.Emails),
-		})
+		match, err := manualDedupePerson(ctx, tx, in)
 		if err != nil {
 			return err
 		}
@@ -140,10 +130,8 @@ func (s *Store) CreatePerson(ctx context.Context, in CreatePersonInput) (crmcont
 		if err := storekit.Emit(ctx, tx, auditID, "person.created", "person", id.UUID, map[string]any{"full_name": in.FullName}); err != nil {
 			return fmt.Errorf("emit person.created: %w", err)
 		}
-		if match.Decision == DecisionFuzzyReview {
-			if err := recordNearMatch(ctx, tx, "person", id.UUID, match.PersonID.UUID, match.Confidence); err != nil {
-				return err
-			}
+		if err := match.recordIfReview(ctx, tx, id); err != nil {
+			return err
 		}
 
 		if out, err = readPerson(ctx, tx, id, storekit.LiveOnly, active); err != nil {
@@ -152,17 +140,6 @@ func (s *Store) CreatePerson(ctx context.Context, in CreatePersonInput) (crmcont
 		return nil
 	})
 	return out, err
-}
-
-// dedupeCandidateEmails flattens the create's child rows into the
-// resolver's input — every address on the candidate counts against
-// PO-F-1, not just the primary.
-func dedupeCandidateEmails(emails []PersonEmailInput) []string {
-	out := make([]string, 0, len(emails))
-	for _, e := range emails {
-		out = append(out, e.Email)
-	}
-	return out
 }
 
 // GetPerson returns one person with child rows; archived rows resolve
@@ -184,133 +161,6 @@ func (s *Store) GetPerson(ctx context.Context, id ids.PersonID, archived storeki
 		return err
 	})
 	return out, err
-}
-
-type ListPeopleInput struct {
-	Cursor          *string
-	Limit           *int
-	Query           *string
-	OwnerID         *ids.UserID
-	IncludeArchived bool
-	// Sort is the contract's sort spec, validated against the core
-	// vocabulary below plus the workspace's active cf_ columns.
-	Sort *string
-	// CustomFilters carries the request's cf_* query parameters —
-	// equality matches against active custom columns (storekit listquery).
-	CustomFilters map[string]string
-}
-
-// personListFields is the person list's core sortable vocabulary —
-// exactly the data-model §13.5 DM-VOCAB-1 set; active cf_ columns join
-// it per request.
-var personListFields = map[string]string{
-	"created_at":  storekit.KindTimestamp,
-	"updated_at":  storekit.KindTimestamp,
-	"full_name":   fieldcatalog.TypeText,
-	ownerIDColumn: storekit.KindUUID,
-}
-
-func (s *Store) ListPeople(ctx context.Context, in ListPeopleInput) ([]crmcontracts.Person, storekit.Page, error) {
-	if err := auth.Require(ctx, "person", principal.ActionRead); err != nil {
-		return nil, storekit.Page{}, err
-	}
-	active, err := s.activeColumns(ctx, "person")
-	if err != nil {
-		return nil, storekit.Page{}, err
-	}
-	sorted, err := storekit.ParseListSort(in.Sort, storekit.SortVocabulary(personListFields, active))
-	if err != nil {
-		return nil, storekit.Page{}, err
-	}
-	limit := storekit.ClampLimit(in.Limit)
-
-	where := []string{"1=1"}
-	args := []any{}
-	arg := func(v any) int { args = append(args, v); return len(args) }
-
-	scope, err := auth.ScopeClauseFor(ctx, "person", "", arg)
-	if err != nil {
-		return nil, storekit.Page{}, err
-	}
-	if scope != "" {
-		where = append(where, scope)
-	}
-
-	if !in.IncludeArchived {
-		where = append(where, "archived_at IS NULL")
-	}
-	if in.OwnerID != nil {
-		where = append(where, storekit.SQLf("owner_id = $%d", arg(*in.OwnerID)))
-	}
-	if in.Query != nil && *in.Query != "" {
-		where = append(where, storekit.QuickFindClause(arg(*in.Query), "full_name"))
-	}
-	cfClauses, err := storekit.CustomFilterClauses(active, in.CustomFilters, arg)
-	if err != nil {
-		return nil, storekit.Page{}, err
-	}
-	where = append(where, cfClauses...)
-	if in.Cursor != nil && *in.Cursor != "" {
-		clause, err := sorted.KeysetClause(*in.Cursor, arg)
-		if err != nil {
-			return nil, storekit.Page{}, err
-		}
-		where = append(where, clause)
-	}
-
-	var people []crmcontracts.Person
-	var page storekit.Page
-	err = s.tx(ctx, func(tx pgx.Tx) error {
-		rows, err := tx.Query(ctx,
-			`SELECT `+personColumns+storekit.SelectSuffix(active)+sorted.CursorKeySuffix()+
-				` FROM person WHERE `+strings.Join(where, " AND ")+
-				sorted.OrderBy()+storekit.SQLf(` LIMIT %d`, limit+1),
-			args...)
-		if err != nil {
-			return err
-		}
-		defer rows.Close()
-
-		var cursorKeys []*string
-		if people, cursorKeys, err = scanPersonPage(rows, active, sorted); err != nil {
-			return err
-		}
-		if len(people) > limit {
-			people = people[:limit]
-			last := people[len(people)-1]
-			page = storekit.Page{HasMore: true, NextCursor: sorted.EncodePageCursor(cursorKeys[limit-1], last.CreatedAt, ids.UUID(last.Id))}
-		}
-		return attachPersonChildren(ctx, tx, people)
-	})
-	if people == nil {
-		people = []crmcontracts.Person{}
-	}
-	return people, page, err
-}
-
-// scanPersonPage drains one list query's rows: each person plus, under a
-// non-default sort, the row's cursor key (the trailing __cursor_key
-// column CursorKeySuffix appended).
-func scanPersonPage(rows pgx.Rows, active []fieldcatalog.Column, sorted *storekit.ListSort) ([]crmcontracts.Person, []*string, error) {
-	var people []crmcontracts.Person
-	var cursorKeys []*string
-	for rows.Next() {
-		var key *string
-		extra := []any{}
-		if sorted != nil {
-			extra = append(extra, &key)
-		}
-		p, err := scanPerson(rows, active, extra...)
-		if err != nil {
-			return nil, nil, err
-		}
-		people = append(people, p)
-		cursorKeys = append(cursorKeys, key)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, nil, err
-	}
-	return people, cursorKeys, nil
 }
 
 type UpdatePersonInput struct {
